+++ contracts/ClearingHouse.sol
@@ -192,8 +192,8 @@ contract ClearingHouse is
         // price slippage check
         require(response.base >= params.minBase && response.quote >= params.minQuote, "CH_PSC");
 
-        // if useTakerPosition, use addBalanceForTaker() instead of addBalance() to modify takerBalances
-        if (params.useTakerPosition) {
+        // if useTakerBalance, use addTakerBalance() instead of addBalance() to modify takerBalances
+        if (params.useTakerBalance) {
             bool isBaseAdded = response.base > 0;
             bool isQuoteAdded = response.quote > 0;
 
@@ -610,7 +610,7 @@ contract ClearingHouse is
     }
 
     function _settleBalanceAndRealizePnl(
-        address trader,
+        address maker,
         address baseToken,
         IOrderBook.RemoveLiquidityResponse memory response
     ) internal {
@@ -618,7 +618,7 @@ contract ClearingHouse is
         if (response.deltaTakerBase != 0) {
             pnlToBeRealized = IExchange(_exchange).getPnlToBeRealized(
                 IExchange.RealizePnlParams({
-                    trader: trader,
+                    trader: maker,
                     baseToken: baseToken,
                     deltaAvailableBase: response.deltaTakerBase,
                     deltaAvailableQuote: response.deltaTakerQuote
@@ -627,7 +627,7 @@ contract ClearingHouse is
         }
 
         IAccountBalance(_accountBalance).settleBalanceAndDeregister(
-            trader,
+            maker,
             baseToken,
             response.base.toInt256(),
             response.quote.toInt256(),