+++ test/clearingHouse/ClearingHouse.accounting.xyk.test.ts
@@ -106,7 +106,7 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
             upperTick,
             minBase: 0,
             minQuote: 0,
-            useTakerPosition: false,
+            useTakerBalance: false,
             deadline: ethers.constants.MaxUint256,
         })