+++ test/clearingHouse/ClearingHouse.closePosition.test.ts
@@ -86,7 +86,7 @@ describe("ClearingHouse closePosition", () => {
                 upperTick, // 151.3733069
                 minBase: 0,
                 minQuote: 0,
-                useTakerPosition: false,
+                useTakerBalance: false,
                 deadline: ethers.constants.MaxUint256,
             }
             await clearingHouse.connect(alice).addLiquidity(addLiquidityParams)
@@ -263,7 +263,7 @@ describe("ClearingHouse closePosition", () => {
                 upperTick, // 151.3733069
                 minBase: 0,
                 minQuote: 0,
-                useTakerPosition: false,
+                useTakerBalance: false,
                 deadline: ethers.constants.MaxUint256,
             }
             // transfer 0.002450462523 base to pool
@@ -278,7 +278,7 @@ describe("ClearingHouse closePosition", () => {
                 upperTick, // 151.3733069
                 minBase: 0,
                 minQuote: 0,
-                useTakerPosition: false,
+                useTakerBalance: false,
                 deadline: ethers.constants.MaxUint256,
             }
             // transfer 0.000816820841 base to pool
@@ -355,7 +355,7 @@ describe("ClearingHouse closePosition", () => {
                     upperTick, // 151.3733069
                     minBase: 0,
                     minQuote: 0,
-                    useTakerPosition: false,
+                    useTakerBalance: false,
                     deadline: ethers.constants.MaxUint256,
                 }
                 // transfer 0.002450462523 base to pool
@@ -370,7 +370,7 @@ describe("ClearingHouse closePosition", () => {
                     upperTick, // 151.3733069
                     minBase: 0,
                     minQuote: 0,
-                    useTakerPosition: false,
+                    useTakerBalance: false,
                     deadline: ethers.constants.MaxUint256,
                 }
                 // transfer 0.000816820841 base to pool
@@ -503,7 +503,7 @@ describe("ClearingHouse closePosition", () => {
                     upperTick, // 151.3733069
                     minBase: 0,
                     minQuote: 0,
-                    useTakerPosition: false,
+                    useTakerBalance: false,
                     deadline: ethers.constants.MaxUint256,
                 }
                 // transfer 0.002450462523 base to pool
@@ -518,7 +518,7 @@ describe("ClearingHouse closePosition", () => {
                     upperTick, // 151.3733069
                     minBase: 0,
                     minQuote: 0,
-                    useTakerPosition: false,
+                    useTakerBalance: false,
                     deadline: ethers.constants.MaxUint256,
                 }
                 // transfer 0.000816820841 base to pool
@@ -639,7 +639,7 @@ describe("ClearingHouse closePosition", () => {
                 upperTick, // 151.3733069
                 minBase: 0,
                 minQuote: 0,
-                useTakerPosition: false,
+                useTakerBalance: false,
                 deadline: ethers.constants.MaxUint256,
             }
             await clearingHouse.connect(alice).addLiquidity(addLiquidityParams)